# api/openapi.yaml
openapi: 3.1.0
info:
  title: RingLink API
  version: 1.0.0
  description: Wrestling talent marketplace MVP (`/api/v1`). Bearer Sanctum personal access tokens.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /wrestlers:
    get:
      security: []
      summary: Search wrestlers
      parameters:
        - in: query
          name: state
          schema:
            type: string
            minLength: 2
            maxLength: 2
        - in: query
          name: wrestling_style
          schema:
            type: string
        - in: query
          name: per_page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
    post:
      summary: Create wrestler profile
      responses:
        '201':
          description: Created
...
  /media/{mediaLink}:
    delete:
      summary: Delete media link
      parameters:
        - name: mediaLink
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum personal access token
...